+++ CHANGELOG.md
@@ -9,13 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * lexemes
-* tab views
+* tabs in morpheme and morph detail views
 
 ### Removed
 * bugs
 
 ### Changed
-* unknown POS are rendered as `*`
+* missing POS are rendered as `*`
 
 ## [0.0.3] - 2022-06-03