+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * missing POS are rendered as `*`
+* wordform structure links to morphemes even in detail view
 
 ## [0.0.3] - 2022-06-03