+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Removed
+
+### Changed
+
+## [0.0.4] -- 2022-06-10
+
+### Added
 * lexemes
 * tabs in morpheme and morph detail views
 
@@ -47,7 +55,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Initial release
 
-[Unreleased]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.3...HEAD
+[Unreleased]: https://github.com/fmatter/clld-morphology-plugin/compare/0.0.4...HEAD
+[0.0.4]: https://github.com/fmatter/clld-morphology-plugin/releases/tag/0.0.4
 [0.0.3]: https://github.com/fmatter/clld-morphology-plugin/releases/tag/0.0.3
 [0.0.2]: https://github.com/fmatter/clld-morphology-plugin/releases/tag/0.0.2
 [0.0.1]: https://github.com/fmatter/clld-morphology-plugin/releases/tag/v0.0.1